0.2beta

+++ backend/database/migrations/2025_08_19_000908_create_orders_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->boolean("is_safe")->default(false);
 
             $table->integer("count_deals")->default(0);
-            $table->integer("remain");
+            $table->decimal("remain", 30, 18);
             $table->string("payment_method")->default("sbp");
 
             $table->integer("price");
